requirements.txt
# vBot 依赖
botpy
uapi-sdk-python>=0.1.0
aiohttp>=3.8.0
pyyaml
